backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.9
spacy==3.7.4
scikit-learn==1.4.2
numpy==1.26.4
pdfplumber==0.11.0
python-docx==1.1.2
reportlab==4.2.0
